requirements.txt
matplotlib==3.10.3
mplfinance==0.12.10b0
numpy==2.3.1
pandas==2.3.1
python_dateutil==2.9.0.post0
yfinance==0.2.65
